Full "airesize" functionality for all ACMEAIR-based workloads - Make sure that the command `airesize <ai name> mongodb +N` results in still functional MongoDB cluster with N additional data nodes - Added a small (AI-specific) utility called `cbcluster`, which will show a snaphsot of the current MongoDB cluster configuration - IMPORTANT: this commit ensure that the cluster will be properly increased in size and still remains functional. It does not deal with data rebalancing (i.e., it is better to perform the AI resize BEFORE start the load generation on the AI)

+++ scripts/mongo_acmeair/cb_restart_mongos.sh
@@ -32,13 +32,26 @@ done
 sudo sed -i "s/bind_ip.*$/bind_ip = ${my_ip_addr}/g" ${MONGODB_CONF_FILE}
 sudo sed -i "s/port.*$/port = 27017/g" ${MONGODB_CONF_FILE}
 
+my_dbpath=$(sudo cat /etc/mongodb.conf | grep dbpath | cut -d '=' -f 2)
+
+sudo ps aux | grep "mongos --configdb cbcsrs/${mongocfg_ip}:27017 --port 27017 --fork --bind_ip 0.0.0.0 --logpath /var/log/mongodb/mongodb.log" | grep -v grep > /dev/null 2>&1
+if [[ $? -ne 0 ]]
+then
+
+    syslog_netcat "Starting mongo router (mongos) on ${SHORT_HOSTNAME}" 
+
+    service_stop_disable mongod
+    service_stop_disable mongodb
+
 #
 # Start Mongos 
 #
-sudo pkill -9 -f configdb
-sudo screen -S MGSS -X quit
-sudo screen -d -m -S MGSS
-sudo screen -p 0 -S MGSS -X stuff "sudo mongos --configdb ${mongocfg_ip}:27019$(printf \\r)"
+    sudo pkill -9 -f configdb
+    sudo screen -S MGSS -X quit
+    sudo screen -d -m -S MGSS
+    sudo screen -p 0 -S MGSS -X stuff "sudo mongos --configdb cbcsrs/${mongocfg_ip}:27017 --port 27017 --fork --bind_ip 0.0.0.0 --logpath /var/log/mongodb/mongodb.log$(printf \\r)"
+    #sudo screen -p 0 -S MGSS -X stuff "sudo mongos --configdb cbcsrs/${mongocfg_ip}:27017 --port 27017 --dbpath ${my_dbpath} --fork --bind_ip 0.0.0.0 --logpath /var/log/mongodb/mongodb.log$(printf \\r)"
+fi
 
 wait_until_port_open ${my_ip_addr} 27017 20 5
 
@@ -54,16 +67,25 @@ fi
 #
 # Add Shards
 #
+
+mongo --host ${mongos_ip}:27017 --eval "db.printShardingStatus()" | grep cbdrs | grep host | awk '{ print $7 }' | sed 's/,//g' | sed 's/"//g' 2>&1 > /tmp/shardlist
+
 pos=1
 for db in $mongo_ips
 do
-    mongo --host ${mongos_ip}:27017 --eval "sh.addShard(\"mongo$pos:27017\")"
-    syslog_netcat " Adding the following shard: mongo$pos:27017 "
+    grep cbdrs${pos}/$db:27017 /tmp/shardlist > /dev/null 2>&1
+    if [[ $? -ne 0 ]]
+    then
+        mongo --host ${mongos_ip}:27017 --eval "sh.addShard(\"cbdrs${pos}/$db:27017\")"
+        syslog_netcat " Adding shard with \"mongo --host ${mongos_ip}:27017 --eval \"sh.addShard(\\"cbdrs${pos}/$db:27017\\")\""
+    fi
     ((pos++))
 done
 
-#mongo --host ${mongos_ip}:27017 --eval "db.printShardingStatus()" | grep \"shard | wc -l
+check_mongodb_cluster_state ${mongos_ip} cbdrs 10 20
+
+STATUS=$?
 
 provision_application_stop $START
 
-exit 0
+exit $STATUS
